Add files via upload
The objective of this assignment was to enhance the visual quality of a set of images by 
applying image beautification techniques. The primary focus was on making the skin appear 
smoother and brighter, using frequency domain techniques combined with additional 
spatial domain enhancements. The images were processed to highlight the improvements 
before and after beautification.